Add first batch of games

+++ nerdvanapp/models/games.py
@@ -9,7 +9,7 @@ class Games(models.Model):
     console = models.ManyToManyField('Console')
     game_company = models.ForeignKey('GameCompany', on_delete=models.CASCADE, blank=True, null=True)
     rating = models.DecimalField(max_digits=5, decimal_places=2, blank=True, null=True)
-    rating_count = models.DecimalField(max_digits=5, decimal_places=2, blank=True, null=True)
+    rating_count = models.IntegerField(blank=True, null=True)
 
     def __str__(self):
         return self.name
